Changes to `seed` in the distribution.
No longer have to specify the key type within the constuction of the
distribution.

Changes to the Gaussian distribution -- instead of having to specify
two keys, we just use the negation of the counter for the second point.
I think it's very doubtful that we end up using negative numbers
as counters...